Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 146ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240821/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -9,7 +9,7 @@ public class Main {
     public static int minCount = 19;
     public static int maxCount = 0;
 
-    public static void find(int target) {
+    public static void find(long target) {
         long left = 1;
         long right = M;
